backend/requirements.txt
Flask
Flask-Cors
Flask-Caching
mistralai
python-dotenv
chromadb
pytest
transformers
torch
sentence-transformers
langchain-core
langchain
